Advent.AoC

Description

Common stuff across AoC solutions.

Parsing

type Parser = Parsec Void Text

A Megaparsec Parser.

parseFile :: Parser a -> String -> IO a

Load a file (e.g. "input/day24") with the given parser.

parseLit :: Parser a -> Text -> a

Parse a literal example.

parseGrid :: (Char -> a) -> String -> Map (Int, Int) a

Parse a grid into a map of (x,y) pairs.

Distances

mdist2 :: (Int, Int) -> (Int, Int) -> Int

Two dimensional manhattan distance.

mdist3 :: (Int, Int, Int) -> (Int, Int, Int) -> Int

Three dimensional manhattan distance.

mdist4 :: (Int, Int, Int, Int) -> (Int, Int, Int, Int) -> Int

Four dimensional manhattan distance.

Tuple Helpers

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thrd :: (a, b, c) -> c

zipt2 :: (a -> b -> c) -> (a, a) -> (b, b) -> (c, c)

Parallel application of a function across elements of a tuple.

zipt3 :: (a -> b -> c) -> (a, a, a) -> (b, b, b) -> (c, c, c)

Parallel application of a function across elements of a tuple.

zipt4 :: (a -> b -> c) -> (a, a, a, a) -> (b, b, b, b) -> (c, c, c, c)

Parallel application of a function across elements of a tuple.

ftup :: (a -> b) -> (a, a) -> (b, b)

fmap for uniform tuples.

ftup3 :: (a -> b) -> (a, a, a) -> (b, b, b)

fmap for uniform tuples

ftup4 :: (a -> b) -> (a, a, a, a) -> (b, b, b, b)

fmap for uniform tuples

zipWith2D :: (x -> y -> a -> r) -> [x] -> [y] -> [[a]] -> [r]

zipWith, but along a two-dimensional input.

Going around in circles

succ' :: (Bounded a, Enum a, Eq a) => a -> a

A circular succ

pred' :: (Bounded a, Enum a, Eq a) => a -> a

A circular pred

ntimes :: Int -> (a -> a) -> a -> a

Repeat an operation n times.

Selection

select :: [a] -> [(a, [a])]

expand a list to all of the variants of the list with an element removed.

select "cat"  ->  [('c',"at"),('a',"ct"),('t',"ca")]

Strange Loops

möb :: (((a -> b) -> b) -> c -> a) -> c -> a

Löb's theorem. https://en.wikipedia.org/wiki/L%C3%B6b%27s_theorem https://github.com/quchen/articles/blob/master/loeb-moeb.md

löb :: Functor f => f (f a -> a) -> f a

Löb's theorem. https://en.wikipedia.org/wiki/L%C3%B6b%27s_theorem https://github.com/quchen/articles/blob/master/loeb-moeb.md