Advent.Grid

Description

This is useful any time you've got a map of x,y coordinates you want to fetch values from without having to put everything into a data structure. Lookups are O(1), storage is approximately the number of bytes of the input, etc...

Documentation

data Grid

Grid represents byte data from a file with equal length lines of newline separated data.

readFile :: FilePath -> IO Grid

Load a Grid from a given file path.

There's not a lot of checking here. Use this on proper grid data.

parseInput :: ByteString -> Grid

Parse a ByteString into a Grid.

There's not a lot of checking here. Use this on proper grid data.

bounds :: Grid -> (Point, Point)

Return the bounds of this Grid.

range :: Grid -> [Point]

Produce a list of all (x,y) Points within this Grid.

index :: Grid -> Point -> Int

Return the numeric index of a point within this Grid.

inRange :: Grid -> Point -> Bool

Return True if the given Point is within this Grid.

lookup :: Grid -> Point -> Maybe Word8

Get the value at the given Point from this Grid iff it's within range.

unsafeLookup :: Grid -> Point -> Word8

Get the value at the given Point from this Grid.

assocs :: Grid -> [(Point, Word8)]

Return all of the Points and values from this Grid.

assocsMap :: (Word8 -> a) -> Grid -> [(Point, a)]

Return all of the Points and values mapped to a particular value. from this Grid.

b2c :: Word8 -> Char

Convert a Word8 to a Char in a fairly unsafe but quick way.

c2b :: Char -> Word8

map :: (Word8 -> Word8) -> Grid -> Grid

Map all the values in the grid.

ixmap :: (Point -> Word8 -> Word8) -> Grid -> Grid

Map all of the values from the grid including the index.

showGrid :: Grid -> String

Show a Grid.

showGridWith :: (Word8 -> Char) -> Grid -> String

Show grid with the given byte -> char mapping.