Advent.Search

Description

Things I use for searching space in AoC.

Graph searching

dijkstra'

Arguments

:: Ord v

=> (v -> [(Int, v)])
Provide a list of all neighbors of v with their associated costs.

-> v
The starting point.

-> (v -> Bool)
Predicate allowing early termination of search (if True).

-> (v, Map v Int, Map v v)
(cost from origin to v, links from points to points)

dijkstra' uses Dijkstra's Algorithm to find the costs and links from a starting point to various points on a graph.

See resolveDijkstra for a means of determining a path from the resulting values.

dijkstra

Arguments

:: Ord v

=> (v -> [(Int, v)])
Provide a list of all neighbors of v with their associated costs.

-> v
The starting point.

-> (v -> Bool)
Goal function

-> Maybe (Int, [v])
The cost to the destination, and the path to get there.

dijkstra uses Dijkstra's Algorithm to find the lowest cost path to a given destination.

resolveDijkstra :: Ord v => Map v Int -> Map v v -> v -> v -> Maybe (Int, [v])

Using maps computed by dijkstra', find the cost and path from the start to a destination.

bfs :: Ord a => (a -> [a]) -> a -> [a]

bfs finds all reachable states from a given value and a function to find its neighbors.

bfsOn

Arguments

:: Ord r

=> (a -> r)
representative function

-> (a -> [a])
neighbors

-> a
initial state

-> [a]
reachable states

bfsOn finds all reachable states from a given value and a function to find its neighbors using a representative function.

The representative function allows a separation between the structure of a state and how it's represented, removing the Ord requirement for the state.

bfsOnInt

Arguments

:: (a -> Int)
representative function

-> (a -> [a])
neighbors

-> a
initial state

-> [a]
reachable states

bfsOnInt finds all reachable states from a given value and a function to find its neighbors using a representative function.

This is a specialization of bfsOn where the representation is an Int.

bfsM

Arguments

:: (Monad m, Ord a)

=> (a -> m [a])
neighbors

-> a
initial state

-> m [a]
reachable states

bfsM finds all reachable states from a given value and a function to find its neighbors.

bfsOnM

Arguments

:: (Monad m, Ord r)

=> (a -> r)
representative function

-> (a -> m [a])
neighbors

-> a
initial state

-> m [a]
reachable states

bfsOn finds all reachable states from a given value and a function to find its neighbors using a representative function.

The representative function allows a separation between the structure of a state and how it's represented, removing the Ord requirement for the state.

Binary searching

binSearch :: Integral a => (a -> Ordering) -> a -> a -> a

binSearch performs a binary search to find the boundary function where a function returns its highest LT value.

autoBinSearch :: Integral a => (a -> Ordering) -> a

A binary search with auto-discovering bounds.

binSearchM :: (Integral a, Monad m) => (a -> m Ordering) -> a -> a -> m a

binSearchM performs a binary search over a monadic action to find the boundary function where a function returns its highest LT value.

List cycle/repeat detection

findCycle :: Ord b => (a -> b) -> [a] -> (Int, Int, a)

Get the position of the start of the first cycle and the cycle length from a list.

findRepeated :: Eq a => [a] -> Maybe a

Get the first repeated element.

findRepeatedOn :: Eq b => (a -> b) -> [a] -> Maybe a

Get the first repeated element using a comparator function

findMin :: Ord b => (a -> b) -> [a] -> a

Find a local minimum.

findMax :: Ord b => (a -> b) -> [a] -> a

Find a local maximum.

countIf :: (a -> Bool) -> [a] -> Int

Count the number of items for which this is true.

Fitting

arranger :: (a -> b -> Bool) -> [a] -> [b] -> Maybe [a]

Find an arrangement of a such that our predicate is happy for each element in b.

e.g.

>>> arranger (<) [1,3,2,4] [3,6,2,6]
Just [2, 3, 1, 4]

List expansion

perturb :: (a -> [a]) -> [a] -> [[a]]

Make variations of a list by changing an element.