Advent.TwoD

Description

Functions for working in 2D space.

Documentation

type Point = (Int, Int)

A point in 2D space.

around :: Point -> [Point]

Get the points around a point (not including diagonals).

aroundD :: Point -> [Point]

Get the points around a point (including diagonals).

data Dir

Directions.

Constructors

N

E

S

W

fwd :: Dir -> Point -> Point

Move a point one position in a particular direction.

fwdBy :: Int -> Dir -> Point -> Point

Move a point some amount in a particular direction.

addPoint :: (Int, Int) -> Point -> Point

Add some to points.

mulPoint :: (Int, Int) -> Point -> Point

Multiply x and y values in a point by the given amounts.